working towards syncing
should probably ditch the seconds_till_next_beat and think of something
better..

updating the position at the end of each frame produced for the players

+++ include/player.hpp
@@ -103,7 +103,7 @@ namespace DataJockey {
 			//internals, bookkeeping, etc
 			unsigned int mSampleRate;
 			float * mVolumeBuffer;
-			unsigned int mSampleIndex;
+			unsigned long mSampleIndex;
 			double mSampleIndexResidual;
 			RubberBand::RubberBandStretcher * mRubberBandStretcher;
 			AudioBuffer * mAudioBuffer;
